# app/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: OpenWatch (Stage 0)
  version: 0.1.0-alpha.1
  description: |
    Stage 0 walking-skeleton surface — three endpoints that exercise the
    foundation stack (correlation, idempotency, audit, error envelope).
servers:
  - url: https://localhost:8443
    description: Local development
paths:
...
  /api/v1/diagnostics:echo:
    post:
      operationId: postDiagnosticsEcho
      summary: Echo the request message; exercises idempotency + audit
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 255}
        - name: X-Correlation-Id
          in: header
          required: false
          schema: {type: string}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EchoRequest'}
      responses:
        '200':
          description: Echo succeeded
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EchoResponse'}
        '400':
          $ref: '#/components/responses/BadRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '409':
          description: Idempotency-Key reused with different body
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  /api/v1/admin/license:verify:
    post:
      operationId: postAdminLicenseVerify
      summary: Dry-run validate a JWT license without installing it
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LicenseVerifyRequest'}
      responses:
        '200':
          description: Verification result (success or typed failure)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LicenseVerifyResponse'}
  /api/v1/diagnostics:premium-echo:
    post:
      operationId: postDiagnosticsPremiumEcho
      summary: License-gated echo; requires premium_diagnostics feature
      x-required-feature: premium_diagnostics
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 255}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EchoRequest'}
      responses:
        '200':
          description: Echoed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EchoResponse'}
        '402':
          description: License does not include premium_diagnostics feature
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/diagnostics:require-host-read:
    post:
      operationId: postDiagnosticsRequireHostRead
      summary: Stage-0 RBAC demo; requires host:read permission
      x-required-permission: host:read
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 255}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EchoRequest'}
      responses:
        '200':
          description: Echoed (caller holds host:read)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EchoResponse'}
        '403':
          description: Caller lacks host:read
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/diagnostics:require-host-write:
    post:
      operationId: postDiagnosticsRequireHostWrite
      summary: Stage-0 RBAC demo; requires host:write permission
      x-required-permission: host:write
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 255}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EchoRequest'}
      responses:
        '200':
          description: Echoed (caller holds host:write)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EchoResponse'}
        '403':
          description: Caller lacks host:write
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/diagnostics:evaluate-alert:
    post:
      operationId: postDiagnosticsEvaluateAlert
      summary: Stage-0 policy demo; evaluates the alert_thresholds policy
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EvaluateAlertRequest'}
      responses:
        '200':
          description: Policy decision
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PolicyDecisionResponse'}
...
  /api/v1/admin/policies:reload:
    post:
      operationId: postAdminPoliciesReload
      summary: Reload signed policy files from the configured directory
      x-required-permission: policy:reload
      responses:
        '200':
          description: Per-type reload outcomes
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PoliciesReloadResponse'}
        '403':
          description: Caller lacks policy:reload
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  /api/v1/diagnostics:require-remediation-execute:
    post:
      operationId: postDiagnosticsRequireRemediationExecute
      summary: Stage-0 RBAC+license demo; requires remediation:execute + remediation_execution feature
      x-required-permission: remediation:execute
      x-required-feature: remediation_execution
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 255}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EchoRequest'}
      responses:
        '200':
          description: Echoed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EchoResponse'}
        '402':
          description: License does not include remediation_execution
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
        '403':
          description: Caller lacks remediation:execute
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  /api/v1/admin/roles:
    get:
      operationId: getAdminRoles
      summary: List built-in roles
      x-required-permission: role:read
      responses:
        '200':
          description: Role list
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminRolesResponse'}
        '403':
          description: Caller lacks role:read
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorEnvelope'}
...
  /api/v1/audit/events:
    get:
      operationId: getAuditEvents
      summary: List audit events (cursor-paginated, newest first)
      parameters:
        - name: action
          in: query
          schema: {type: string}
        - name: correlation_id
          in: query
          schema: {type: string}
        - name: actor_type
          in: query
          schema: {type: string}
        - name: since
          in: query
          schema: {type: string, format: date-time}
        - name: until
          in: query
          schema: {type: string, format: date-time}
        - name: cursor
          in: query
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: Page of audit events
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuditEventsPage'}
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  responses:
    BadRequest:
      description: Validation, pagination, or idempotency error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
    MethodNotAllowed:
      description: Wrong HTTP method
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
...
    EchoRequest:
      type: object
      required: [message]
      properties:
        message: {type: string, minLength: 1, maxLength: 1024}
    EchoResponse:
      type: object
      required: [echoed, correlation_id]
      properties:
        echoed: {type: string}
        correlation_id: {type: string}
        audit_event_id: {type: string, format: uuid}
...
    LicenseVerifyRequest:
      type: object
      required: [license_jwt]
      properties:
        license_jwt: {type: string}
    LicenseVerifyResponse:
      type: object
      required: [is_valid, verify_result]
      properties:
        is_valid: {type: boolean}
        verify_result: {type: string}
        tier: {type: string, nullable: true}
        features: {type: array, items: {type: string}, nullable: true}
        expires_at: {type: string, format: date-time, nullable: true}
        warnings: {type: array, items: {type: string}, nullable: true}
    AuditEvent:
      type: object
      required: [id, correlation_id, action, actor_type, occurred_at]
      properties:
        id: {type: string, format: uuid}
        correlation_id: {type: string}
        action: {type: string}
        severity: {type: string}
        actor_type: {type: string}
        actor_id: {type: string, nullable: true}
        resource_type: {type: string, nullable: true}
        resource_id: {type: string, nullable: true}
        occurred_at: {type: string, format: date-time}
        recorded_at: {type: string, format: date-time}
        detail: {type: object, additionalProperties: true}
        redactions: {type: array, items: {type: string}}
    AuditEventsPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/AuditEvent'}
        next_cursor:
          type: string
          nullable: true
          description: Opaque cursor; pass as ?cursor= to fetch the next page
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, fault, retryable, human_message]
          properties:
            code: {type: string}
            fault: {type: string, enum: [client, server, policy, external]}
            retryable: {type: boolean}
            human_message: {type: string}
            detail: {type: object, additionalProperties: true}
            correlation_id: {type: string}
...
    RoleEntry:
      type: object
      required: [id, description, permissions, is_built_in]
      properties:
        id: {type: string}
        description: {type: string}
        is_built_in: {type: boolean}
        permissions:
          type: array
          items: {type: string}
...
    AdminRolesResponse:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          items: {$ref: '#/components/schemas/RoleEntry'}
    EvaluateAlertRequest:
      type: object
      required: [score]
      properties:
        score: {type: integer, minimum: 0, maximum: 100}
    PolicyDecisionResponse:
      type: object
      required: [outcome, policy_type, policy_version, reason]
      properties:
        outcome:        {type: string}
        policy_type:    {type: string}
        policy_version: {type: string}
        reason:         {type: string}
        human_message:  {type: string}
        detail:         {type: object, additionalProperties: true}
...
    PoliciesReloadResponse:
      type: object
      required: [outcomes]
      properties:
        outcomes:
          type: object
          additionalProperties: {type: string}